feat: add user logic

+++ src/user/user.controller.ts
@@ -1,7 +1,18 @@
-import { Controller, Delete, Get, Post, Put } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  HttpException,
+  HttpStatus,
+  Param,
+  Post,
+  Put,
+} from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
-import { UpdatePasswordDto } from './dto/update-user.dto';
+import { UpdatePasswordDto } from './dto/update-password.dto';
+import { isUUID } from 'class-validator';
 
 @Controller('user')
 export class UserController {
@@ -13,22 +24,74 @@ export class UserController {
   }
 
   @Get(':id')
-  async getUserById(id: string) {
-    return this.userService.getUserById(id);
+  async getUserById(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new HttpException(
+        'Invalid user ID (not UUID)',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    const user = await this.userService.getUserById(id);
+
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return user;
   }
 
-  @Post('create')
-  async createUser(createUserDto: CreateUserDto) {
+  @Post()
+  async createUser(@Body() createUserDto: CreateUserDto) {
     return this.userService.create(createUserDto);
   }
 
   @Put(':id')
-  async updateUser(id: string, updatePasswordDto: UpdatePasswordDto) {
-    return this.userService.updatePassword(id, updatePasswordDto);
+  async updatePassword(
+    @Param('id') id: string,
+    @Body() updatePasswordDto: UpdatePasswordDto,
+  ) {
+    if (!isUUID(id)) {
+      throw new HttpException(
+        'Invalid user ID (not UUID)',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    try {
+      const user = await this.userService.updatePassword(id, updatePasswordDto);
+      if (!user) {
+        throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+      }
+      return user;
+    } catch (error) {
+      if (
+        error instanceof Error &&
+        error.message === 'Old password is incorrect'
+      ) {
+        throw new HttpException(
+          'Old password is incorrect',
+          HttpStatus.FORBIDDEN,
+        );
+      }
+      throw error;
+    }
   }
 
   @Delete(':id')
-  async deleteUser(id: string) {
-    return this.userService.remove(id);
+  async deleteUser(@Param('id') id: string) {
+    if (!isUUID(id)) {
+      throw new HttpException(
+        'Invalid user ID (not UUID)',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    const deleted = await this.userService.remove(id);
+    if (!deleted) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    throw new HttpException('', HttpStatus.NO_CONTENT);
   }
 }
